// fmea-web-client/src/libs/api-client/model/teamMemberDto.ts
/**
 * Fmea API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TeamMemberDto { 
    name?: string | null;
    employeeNo?: string | null;
    role?: string | null;
    department?: string | null;
    email?: string | null;
    phone?: string | null;
    note?: string | null;
}